fix: Wrong `CommitSessionAction`
