+++ src/ai/backend/manager/services/vfolder/processors/base.py
@@ -9,6 +9,8 @@ from ..actions.base import (
     DeleteForeverVFolderActionResult,
     ForceDeleteVFolderAction,
     ForceDeleteVFolderActionResult,
+    GetTaskLogsAction,
+    GetTaskLogsActionResult,
     GetVFolderAction,
     GetVFolderActionResult,
     ListVFolderAction,
@@ -39,6 +41,7 @@ class VFolderBaseProcessors:
     ]
     force_delete_vfolder: ActionProcessor[ForceDeleteVFolderAction, ForceDeleteVFolderActionResult]
     clone_vfolder: ActionProcessor[CloneVFolderAction, CloneVFolderActionResult]
+    get_task_logs: ActionProcessor[GetTaskLogsAction, GetTaskLogsActionResult]
 
     def __init__(self, service: VFolderService):
         self.create_vfolder = ActionProcessor(service.create)
@@ -50,3 +53,4 @@ class VFolderBaseProcessors:
         self.delete_forever_vfolder = ActionProcessor(service.delete_forever)
         self.force_delete_vfolder = ActionProcessor(service.force_delete)
         self.clone_vfolder = ActionProcessor(service.clone)
+        self.get_task_logs = ActionProcessor(service.get_task_logs)
